python-alternate/requirements.txt
# Flask web application
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Database
SQLAlchemy==2.0.23
alembic==1.13.0

# Discord integration
discord.py==2.3.2
requests==2.31.0

# OSRS monitoring
Pillow==10.1.0
pytesseract==0.3.10
pyautogui==0.9.54
watchdog==3.0.0

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Web server (production)
gunicorn==21.2.0
waitress==2.1.2

# Caching (optional)
redis==5.0.1
Flask-Caching==2.1.0

# Security
bcrypt==4.1.1
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
Jinja2==3.1.2
MarkupSafe==2.1.3
